Moving sign out and change password
